arm/modifyreg: Use common lock to protect the modify address to ensure consistency
