+++ arch/arm/src/common/Make.defs
@@ -24,9 +24,8 @@
 
 CMN_CSRCS += arm_allocateheap.c arm_createstack.c arm_exit.c
 CMN_CSRCS += arm_getintstack.c arm_initialize.c arm_lowputs.c
-CMN_CSRCS += arm_modifyreg8.c arm_modifyreg16.c arm_modifyreg32.c
 CMN_CSRCS += arm_nputs.c arm_releasestack.c arm_registerdump.c
-CMN_CSRCS += arm_stackframe.c
+CMN_CSRCS += arm_stackframe.c arm_modifyreg.c
 CMN_CSRCS += arm_usestack.c arm_fork.c
 
 ifneq ($(CONFIG_ALARM_ARCH),y)
